Update nsftool.py
* Added Command line options.
* Added two new functions: decrypt(-d),  crack or brute force(-c)

Decrypt function clones and compacts NSF's that past the valid check.

Crack or the brute force function tries every combinations of ID file Password combination found in the load file to try and brute force the nsf.   I thought this is useful given our clients, nothing else to say here.

Notes:
Steve's method seemed to not work when dealing with multiple ID files.  For some reason LN is only reading the id file listed in the INI file.  For this reason I have cleaned-up and reverted back to the first ID method.